pulsar-client-py/requirements.txt
numpy==1.23.4
pandas==1.5.0
pulsar-client==2.10.1
